Move Task constructor and const field qualifiers to config.js. Add dependencies as bundled.
