+++ .npmignore
@@ -1,0 +1,2 @@
+*.tar.gz
+*.sh
