+++ .tasks/data/4543a31eb17bc37acece55c65d3e6136a9e5f0636acce092c072340e889a9026
@@ -1,0 +1,11 @@
+{
+  "status": "done",
+  "pending": "0",
+  "estimation": 1,
+  "title": "Add the new dependencies as bundled",
+  "tags": "package",
+  "gitStatus": "A ",
+  "id": "4543a31eb17bc37acece55c65d3e6136a9e5f0636acce092c072340e889a9026",
+  "createdAt": 1474339096110,
+  "updatedAt": 1474339142072
+}
