+++ lib/sync/jira-client.js
@@ -5,6 +5,7 @@
 require('../common');
 var argv = require('../argv');
 var JiraApi = require('jira-client');
+var server = require('../server');
 
 const CLI_OPTIONS = [
     {
@@ -71,6 +72,15 @@ JiraClient.prototype.getIssues = function(projectName, callback, errorCallback) 
         errorCallback(error);
     });
 };
+
+//TODO: Complete array definition
+const LOCAL_TASK_MAP_FIELDS = [
+    {
+        status: function (issue) {
+            return issue['fields']['Name'];
+        }
+    }
+];
 
 /**
  * Get git-task mapped issues.
