+++ package.json
@@ -20,9 +20,9 @@
     "issue trackers",
     "synchronization"
   ],
-  "dependencies": {
-    "jira-client": "3.0.3"
-  },
+  "bundledDependencies": [
+    "jira-client"
+  ],
   "author": "Jeffrey Alvarez Massón <kuritsu@gmail.com>",
   "license": "MIT"
 }
