+++ site/tasks_dashboard.js
@@ -4,8 +4,6 @@
 
 var tasks = [];
 var draggedTask = null;
-const READONLY_FIELDS = ["id", "createdAt", "updatedAt", "gitStatus", "status"];
-const DATE_FIELDS = ["createdAt", "updatedAt"];
 
 /**
  * Finds the container associated with an status.
@@ -15,19 +13,6 @@ const DATE_FIELDS = ["createdAt", "updatedAt"];
 function getContainerFor(status) {
     var result = $('.column_task_group[data-status="' + status + '"]');
     return result;
-}
-
-/**
- * Task class
- * @constructor
- */
-function Task() {
-    this.status = 'open';
-    this.pending = 1;
-    this.estimation = 1;
-    this.title = 'New task';
-    this.tags = '';
-    this.gitStatus = 'A ';
 }
 
 /**
